v2python: Use SQLite tuning database to direct the build process
